refactor(cli): extract shared utilities, fix security and robustness issues
- Extract runCloudLogin to cli/src/utils/cloud-login.ts (shared by
  login.ts and setup-skills.ts, removes 50 lines of duplication)
- Extract runCliAsync to cli/tests/helpers/cli.ts (shared test helper)
- Fix command injection in browser.ts: replace exec() with execFile()
- Add per-request timeouts to auth.ts: 30s for authorize, 15s for polls
- Guard .json() against non-JSON error responses in auth.ts
- Fix config mutation in open.ts: shallow-copy before modifying
- Remove unused mock server dead code from login.test.ts
- Rename _minIntervalMs to minIntervalMs, fix slow_down cap
