+++ cli/src/utils/browser.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2024 Ivan Murzak. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
 
-import { exec } from 'child_process';
+import { execFile } from 'child_process';
 import { verbose } from './ui.js';
 
 /**
@@ -10,18 +10,22 @@ import { verbose } from './ui.js';
  */
 export function openBrowser(url: string): void {
   const platform = process.platform;
-  let command: string;
+  let cmd: string;
+  let args: string[];
 
   if (platform === 'darwin') {
-    command = `open "${url}"`;
+    cmd = 'open';
+    args = [url];
   } else if (platform === 'win32') {
-    command = `start "" "${url}"`;
+    cmd = 'cmd';
+    args = ['/c', 'start', '', url];
   } else {
-    command = `xdg-open "${url}"`;
+    cmd = 'xdg-open';
+    args = [url];
   }
 
-  verbose(`Opening browser: ${command}`);
-  exec(command, (err) => {
+  verbose(`Opening browser: ${cmd} ${args.join(' ')}`);
+  execFile(cmd, args, (err) => {
     if (err) {
       verbose(`Failed to open browser: ${err.message}`);
     }
