+++ cli/tests/setup-skills.test.ts
@@ -7,11 +7,7 @@ import * as net from 'net';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as os from 'os';
-import { spawn } from 'child_process';
-import { fileURLToPath } from 'url';
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const CLI_PATH = path.resolve(__dirname, '..', 'bin', 'unity-mcp-cli.js');
+import { runCliAsync } from './helpers/cli.js';
 
 // ---------------------------------------------------------------------------
 // Helpers
@@ -64,39 +60,6 @@ function startCaptureServer(): Promise<CaptureServer> {
         },
         close: () => new Promise<void>((res) => server.close(() => res())),
       });
-    });
-  });
-}
-
-/** Run the CLI as a child process and collect stdout+stderr. */
-function runCliAsync(args: string[]): Promise<{ stdout: string; exitCode: number }> {
-  return new Promise((resolve) => {
-    const child = spawn('node', [CLI_PATH, ...args], { stdio: 'pipe' });
-    let stdout = '';
-    let settled = false;
-    const timeoutMs = 30000;
-
-    const timeout = setTimeout(() => {
-      if (settled) return;
-      settled = true;
-      try { child.kill(); } catch { /* noop */ }
-      stdout += '\n[runCliAsync] Process timed out.\n';
-      resolve({ stdout, exitCode: 1 });
-    }, timeoutMs);
-
-    const finish = (exitCode: number) => {
-      if (settled) return;
-      settled = true;
-      clearTimeout(timeout);
-      resolve({ stdout, exitCode });
-    };
-
-    child.stdout?.on('data', (d: Buffer) => { stdout += d.toString(); });
-    child.stderr?.on('data', (d: Buffer) => { stdout += d.toString(); });
-    child.on('close', (code) => { finish(code ?? 0); });
-    child.on('error', (err) => {
-      stdout += `\n[runCliAsync] Error: ${String(err)}\n`;
-      finish(1);
     });
   });
 }
